Move common code to common crate
