+++ Code/Cargo.toml
@@ -1,4 +1,7 @@
 [workspace]
 resolver = "2"
 
-members = []
+members = [
+  "common",
+  "round-sm"
+]
