+++ Code/common/Cargo.toml
@@ -1,9 +1,8 @@
 [package]
-name    = "round-sm"
+name = "common"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-common = { version = "0.1.0", path = "../common" }
